+++ aniso8601/decimalfraction.py
@@ -6,10 +6,6 @@
 # This software may be modified and distributed under the terms
 # of the BSD license.  See the LICENSE file for details.
 
-def find_separator(value):
-    """Returns the decimal separator index if found else -1."""
-    return normalize(value).find('.')
-
 def normalize(value):
     """Returns the string with decimal separators normalized."""
     return value.replace(',', '.')